Handling errors the proper way
Now, if there is an error on the visualize.js side when rendering the input controls, we will trigger the error callback in the new input controls package.
Additionally, when creating the root context in react, if this action fails, we will also trigger the error callback in the new input controls package.